re-enable destroy() tests
